+++ Source/default.py
@@ -9,6 +9,7 @@ import filter
 import gui
 import unitTests
 import time
+import profileGUI
 
 searchProfile.install()
 
@@ -28,4 +29,11 @@ else:
     xbmc.log("Unit Testing Failed!", 0)
 '''
 
-gui.showGui()
+# gui.showGui()
+
+dialog = xbmcgui.Dialog()
+result = dialog.contextmenu(['New Search', 'Saved Search'])
+if (result == 0):
+    gui.showGui()
+elif (result == 1):
+    profileGUI.SavedSearch.showGui()